+++ r/R/kernel.R
@@ -12,7 +12,8 @@
 #' @param feature_id Numeric/Factor, column in data specifying the feature.
 #' @param value_id Numeric, column in data specifying the feature value.
 #' @param response_id Numeric, column in data specifying the response.
-#' @param method Character, the kernel computation method, either "difference", "glm", or "glmm".
+#' @param method Character, the kernel computation method, either "difference", "GLM", or "GLMM".
+#' @param max_feature_levels Numeric, maximum number of feature levels in the GLMM (keep it low to keep GLMM fitting feasible).
 #' @param double_pass Logical, indicating whether the last block was repeated.
 #'
 #' @return The original data augmented with the kernel.
@@ -27,11 +28,16 @@
 #' head(self_voice)
 #'
 #' # computing the kernel per participant
-#' computing_kernel(data = self_voice, method = "difference")
+#' computing_kernel(data = self_voice, method = "difference") |> head()
+#' computing_kernel(data = self_voice, method = "GLM") |> head()
 #'
 #' # plotting it
 #' computing_kernel(data = self_voice, method = "difference") |> plot()
-#' computing_kernel(data = self_voice, method = "glm") |> plot()
+#' computing_kernel(data = self_voice, method = "GLM") |> plot()
+#'
+#' # computing the kernel using a Bayesian multilevel GLM
+#' # note that this option is much slower...
+#' kernel_glmm <- computing_kernel(data = self_voice, method = "GLMM")
 #' }
 #'
 #' @author Ladislas Nalborczyk \email{ladislas.nalborczyk@@gmail.com}.
@@ -42,7 +48,8 @@ computing_kernel <- function (
         data,
         participant_id = "participant", block_id = "block", trial_id = "trial",
         feature_id = "feature", value_id = "value", response_id = "response",
-        method = c("difference", "glm", "glmm"),
+        method = c("difference", "GLM", "GLMM"),
+        max_feature_levels = 5,
         double_pass = TRUE
         ) {
 
@@ -71,7 +78,7 @@ computing_kernel <- function (
 
         # first-order kernel computed for each subject as
         # mean(filter gains of voices classified as mine) - mean(filter gains of voices classified as not-mine)
-        kernel <- data |>
+        kernels <- data |>
             # grouping by participant, frequency band, and response
             dplyr::group_by(.data[[participant_id]], .data[[feature_id]], .data[[response_id]]) |>
             # computing the average gain
@@ -95,9 +102,9 @@ computing_kernel <- function (
             dplyr::ungroup()
 
         # setting the class of the resulting object
-        class(kernel) <- c("kernel", "data.frame")
+        class(kernels) <- c("kernel", "data.frame")
 
-    } else if (method == "glm") {
+    } else if (method == "GLM") {
 
         # constructing the formula dynamically
         formula_str <- paste(response_id, "~ 1 +", value_id, "*", feature_id)
@@ -135,7 +142,7 @@ computing_kernel <- function (
         ci_upper <- as.numeric(c(reference_ci[2], reference_ci[2] + interaction_cis[, 2]) )
 
         # creating a dataframe with results
-        kernel <- data.frame(
+        kernels <- data.frame(
             feature = feature_levels,
             kernel = slopes,
             lower = ci_lower,
@@ -143,159 +150,253 @@ computing_kernel <- function (
             )
 
         # setting the class of the resulting object
-        class(kernel) <- c("glm_kernel", "data.frame")
+        class(kernels) <- c("glm_kernel", "data.frame")
 
-    } else if (method == "glmm") {
-
-        # reshaping the data
-        data2 <- data |>
-            # removing the double-pass block (if needed)
-            dplyr::filter(if (double_pass) .data[[block_id]] < max(.data[[block_id]]) else TRUE) |>
-            dplyr::mutate(feature = as.factor(.data[[feature_id]]) ) |>
-            dplyr::mutate(participant = as.factor(.data[[participant_id]]) ) |>
-            # converting to data table
-            data.table::as.data.table()
-
-        # fitting a multilevel generalised (binomial) linear model per participant
-        # this does not work for the moment...
-        # model <- lme4::glmer(
-        #     formula = resp ~ 1 + eq * f + (1 + eq | participant) + (1 + eq | participant:f),
-        #     family = stats::binomial(),
-        #     data = data2
-        #     )
-        # model2 <- lme4::glmer(
-        #     formula = resp ~ 1 + eq * f + (1 + eq * f | participant),
-        #     family = stats::binomial(),
-        #     data = data |> dplyr::mutate(f = as.factor(.data$f) )
-        #     )
-
-        # constructing the formula dynamically
-        formula_str <- paste(response_id, "~ 1 +", value_id, "*", feature_id)
-        formula_obj <- as.formula(formula_str)
-
-        # using a GAM instead
-        # see https://m-clark.github.io/posts/2019-10-20-big-mixed-models/
-        gamm <- mgcv::bam(
-            formula = resp ~ 1 + eq * f +
-                s(participant, bs = "re") +
-                s(eq, f, participant, bs = "re"),
-                # s(participant, by = eq, bs = "re"),
-                # s(participant, by = f, bs = "re") +
-                # s(participant, by = interaction(eq, f), bs = "re"),
-            family = stats::binomial(link = "logit"),
-            data = data2,
-            discrete = TRUE,
-            # https://stackoverflow.com/questions/78590558/how-can-i-enable-openmp-in-mac-os-for-mgcv-r-package
-            nthreads = parallel::detectCores() - 1,
-            method = "fREML"
-            )
-
-        ############################
-        # retrieving fixed effects #
-        ############################
-
-        # retrieving model coefficients
-        coefficients <- stats::coef(gamm)
-
-        # main effect of eq
-        eq_coef <- coefficients["eq"]
-
-        # interaction terms
-        eq_f_interaction <- coefficients[grepl("^eq:f", names(coefficients) )]
-
-        # creating a table of slopes for eq at each f
-        group_kernel <- data.frame(
-            f = sub("eq:f", "", names(eq_f_interaction) ),
-            kernel_gain = eq_coef + eq_f_interaction
-            ) |>
-            dplyr::mutate(participant = "group") |>
-            dplyr::relocate(.data$participant, 1)
-
-        # removing rownames
-        rownames(group_kernel) <- NULL
-
-        # generating a grid of possible values for predictions
-        # new_data <- expand.grid(
-        #     eq = unique(data2$eq),
-        #     f = unique(data2$f),
-        #     participant = NA
-        #     )
-
-        # retrieving predictions with SE
-        # new_data$predicted <- predict(
-        #     gamm, newdata = new_data, type = "link",
-        #     exclude = "s(participant)",
-        #     se.fit = TRUE
-        #     )
-
-        #############################
-        # retrieving random effects #
-        #############################
-
-        # creating dataset with small variations in eq to estimate slopes
-        new_data <- tidyr::crossing(
-            participant = unique(data2$participant),
-            f = unique(data2$f)
-            ) |>
-            dplyr::mutate(
-                # small perturbation below mean eq
-                eq_low = mean(data2$eq) - 0.01,
-                # small perturbation above mean eq
-                eq_high = mean(data2$eq) + 0.01
-                )
-
-        # predicting values at slightly different eq levels
-        new_data$pred_low <- stats::predict(
-            gamm, newdata = dplyr::mutate(new_data, eq = .data$eq_low),
-            type = "link"
-            )
-        new_data$pred_high <- stats::predict(
-            gamm, newdata = dplyr::mutate(new_data, eq = .data$eq_high),
-            type = "link"
-            )
-
-        # computing approximate slope (kernel) as finite difference
-        kernel <- new_data |>
-            dplyr::mutate(
-                kernel_gain = (.data$pred_high - .data$pred_low) /
-                    (.data$eq_high - .data$eq_low)
-                ) |>
-            dplyr::select(.data$participant, .data$f, .data$kernel_gain)
-
-        # appending fixed and random effects
-        kernel <- dplyr::bind_rows(kernel, group_kernel)
-
-        # setting the class of the resulting object
-        class(kernel) <- c("glmm_kernel", "data.frame")
-
-    # } else if (method == "bglmm") {
+    # } else if (method == "glmm") {
     #
     #     # reshaping the data
     #     data2 <- data |>
-    #         # focusing on speech-relevant frequency bands
-    #         dplyr::filter(.data$f > 100) |>
-    #         dplyr::filter(.data$f < 10000) |>
     #         # removing the double-pass block (if needed)
-    #         # {if (double_pass) dplyr::filter(., .data$block < max(.data$block) ) else .} |>
-    #         dplyr::filter(if (double_pass) .data$block < max(.data$block) else TRUE) |>
-    #         dplyr::mutate(f = as.factor(.data$f) ) |>
-    #         dplyr::mutate(participant = as.factor(.data$participant) ) |>
+    #         dplyr::filter(if (double_pass) .data[[block_id]] < max(.data[[block_id]]) else TRUE) |>
+    #         dplyr::mutate(feature = as.factor(.data[[feature_id]]) ) |>
+    #         dplyr::mutate(participant = as.factor(.data[[participant_id]]) ) |>
     #         # converting to data table
     #         data.table::as.data.table()
     #
-    #     # Bayesian GLMM approach with participant as varying effect
-    #     bglmm <- brms::brm(
-    #         formula = resp ~ 1 + eq * f + (1 | participant),
-    #         family = brms::bernoulli(),
-    #         data = data2,
-    #         chains = 4,
-    #         cores = 4
+    #     # fitting a multilevel generalised (binomial) linear model per participant
+    #     # this does not work for the moment...
+    #     # model <- lme4::glmer(
+    #     #     formula = resp ~ 1 + eq * f + (1 + eq | participant) + (1 + eq | participant:f),
+    #     #     family = stats::binomial(),
+    #     #     data = data2
+    #     #     )
+    #     # model2 <- lme4::glmer(
+    #     #     formula = resp ~ 1 + eq * f + (1 + eq * f | participant),
+    #     #     family = stats::binomial(),
+    #     #     data = data |> dplyr::mutate(f = as.factor(.data$f) )
+    #     #     )
+    #
+    #     # constructing the formula dynamically
+    #     formula_str <- paste(
+    #         response_id, "~ 1 +", value_id, "*", feature_id,
+    #         "+ s(", participant_id, ", bs = 're')",
+    #         "+ s(", value_id, ",", feature_id, ",", participant_id, ", bs = 're')"
     #         )
+    #     formula_obj <- as.formula(formula_str)
+    #
+    #     # using a GAM instead
+    #     # see https://m-clark.github.io/posts/2019-10-20-big-mixed-models/
+    #     gamm <- mgcv::bam(
+    #         formula = formula_obj,
+    #             # old formula (working)
+    #             # resp ~ 1 + eq * f +
+    #             # s(participant, bs = "re") +
+    #             # s(eq, f, participant, bs = "re"),
+    #             # old formula (not working)
+    #             # s(participant, by = eq, bs = "re"),
+    #             # s(participant, by = f, bs = "re") +
+    #             # s(participant, by = interaction(eq, f), bs = "re"),
+    #         family = stats::binomial(link = "logit"),
+    #         data = data2,
+    #         discrete = TRUE,
+    #         # https://stackoverflow.com/questions/78590558/how-can-i-enable-openmp-in-mac-os-for-mgcv-r-package
+    #         nthreads = parallel::detectCores() - 1,
+    #         method = "fREML"
+    #         )
+    #
+    #     ############################
+    #     # retrieving fixed effects #
+    #     ############################
+    #
+    #     # retrieving model coefficients
+    #     coefficients <- stats::coef(gamm)
+    #
+    #     # main effect of value
+    #     value_coef <- coefficients[value_id]
+    #
+    #     # interaction terms
+    #     value_feauture_interaction <- coefficients[grepl("^value:feature", names(coefficients) )]
+    #
+    #     # creating a table of slopes for eq at each f
+    #     group_kernel <- data.frame(
+    #         feature = sub("value:feature", "", names(value_feauture_interaction) ),
+    #         kernel_gain = value_coef + value_feauture_interaction
+    #         ) |>
+    #         dplyr::mutate(participant = "group") |>
+    #         dplyr::relocate(.data[[participant_id]], 1)
+    #
+    #     # removing rownames
+    #     rownames(group_kernel) <- NULL
+    #
+    #     # generating a grid of possible values for predictions
+    #     # new_data <- expand.grid(
+    #     #     eq = unique(data2$eq),
+    #     #     f = unique(data2$f),
+    #     #     participant = NA
+    #     #     )
+    #
+    #     # retrieving predictions with SE
+    #     # new_data$predicted <- predict(
+    #     #     gamm, newdata = new_data, type = "link",
+    #     #     exclude = "s(participant)",
+    #     #     se.fit = TRUE
+    #     #     )
+    #
+    #     #############################
+    #     # retrieving random effects #
+    #     #############################
+    #
+    #     # creating dataset with small variations in eq to estimate slopes
+    #     new_data <- tidyr::crossing(
+    #         participant = unique(data2$participant),
+    #         f = unique(data2$f)
+    #         ) |>
+    #         dplyr::mutate(
+    #             # small perturbation below mean eq
+    #             eq_low = mean(data2$eq) - 0.01,
+    #             # small perturbation above mean eq
+    #             eq_high = mean(data2$eq) + 0.01
+    #             )
+    #
+    #     # predicting values at slightly different eq levels
+    #     new_data$pred_low <- stats::predict(
+    #         gamm, newdata = dplyr::mutate(new_data, eq = .data$eq_low),
+    #         type = "link"
+    #         )
+    #     new_data$pred_high <- stats::predict(
+    #         gamm, newdata = dplyr::mutate(new_data, eq = .data$eq_high),
+    #         type = "link"
+    #         )
+    #
+    #     # computing approximate slope (kernel) as finite difference
+    #     kernel <- new_data |>
+    #         dplyr::mutate(
+    #             kernel_gain = (.data$pred_high - .data$pred_low) /
+    #                 (.data$eq_high - .data$eq_low)
+    #             ) |>
+    #         dplyr::select(.data$participant, .data$f, .data$kernel_gain)
+    #
+    #     # appending fixed and random effects
+    #     kernel <- dplyr::bind_rows(kernel, group_kernel)
+    #
+    #     # setting the class of the resulting object
+    #     class(kernel) <- c("glmm_kernel", "data.frame")
+
+    } else if (method == "GLMM") {
+
+        # retrieving the unique feature levels
+        unique_feature_levels <- unique(data[[feature_id]])
+
+        if (length(unique_feature_levels <- unique(data[[feature_id]]) ) > max_feature_levels) {
+
+            # subsetting the feature levels
+            unique_feature_levels <- unique_feature_levels[1:max_feature_levels]
+
+            # warning the use that we did subset the data
+            cat(
+            "Beware that we only retained the first", max_feature_levels, "feature levels to maintain a reasonable fitting time...\n"
+            )
+
+            # reshaping the data
+            data2 <- data |>
+                # removing the double-pass block (if needed)
+                dplyr::filter(if (double_pass) .data[[block_id]] < max(.data[[block_id]]) else TRUE) |>
+                # reshaping the original variables
+                dplyr::mutate(feature = as.factor(.data[[feature_id]]) ) |>
+                dplyr::mutate(participant = .data[[participant_id]]) |>
+                dplyr::mutate(response = .data[[response_id]]) |>
+                dplyr::mutate(value = .data[[value_id]]) |>
+                # removing some features to keep the fitting reasonable
+                dplyr::filter(.data$feature %in% unique_feature_levels) |>
+                # removing potential NAs
+                stats::na.omit()
+
+        } else {
+
+            # reshaping the data
+            data2 <- data |>
+                # removing the double-pass block (if needed)
+                dplyr::filter(if (double_pass) .data[[block_id]] < max(.data[[block_id]]) else TRUE) |>
+                # reshaping the original variables
+                dplyr::mutate(feature = as.factor(.data[[feature_id]]) ) |>
+                dplyr::mutate(participant = .data[[participant_id]]) |>
+                dplyr::mutate(response = .data[[response_id]]) |>
+                dplyr::mutate(value = .data[[value_id]]) |>
+                # removing potential NAs
+                stats::na.omit()
+
+        }
+
+        # Bayesian GLMM approach with varying intercept and slope per participant
+        glmm <- brms::brm(
+            # formula = response ~ 1 + value * feature + (1 + value * feature | participant),
+            # formula = response ~ 1 + value * feature + (1 + value * feature || participant),
+            formula = response ~ 0 + value * feature + (0 + value * feature || participant),
+            family = brms::bernoulli(),
+            data = data2,
+            # prior = c(
+            #     brms::set_prior(prior = "normal(0, 1)", class = "b"),
+            #     brms::set_prior(prior = "exponential(1)", class = "sd")
+            #     ),
+            warmup = 1000,
+            iter = 2000,
+            chains = 4,
+            cores = 4
+            )
+
+        # retrieving the group-level (fixed/constant) effects
+        fixed_effects <- brms::fixef(glmm)
+        group_kernel <- fixed_effects |>
+            as.data.frame() |>
+            tibble::rownames_to_column(var = "variable") |>
+            dplyr::filter(grepl("value", .data$variable) ) |>
+            dplyr::mutate(participant = "group", .before = .data$variable) |>
+            dplyr::mutate(variable =  sub("\\..*", "", .data$variable) ) |>
+            dplyr::mutate(variable =  sub(".*:", "", .data$variable) )
+
+        # reshaping to provide the group-level kernel
+        group_kernel <- data.frame(
+            feature = group_kernel$variable,
+            kernel = group_kernel$Estimate,
+            lower = group_kernel$Q2.5,
+            upper = group_kernel$Q97.5
+            ) |>
+            dplyr::mutate(participant = "group", .before = .data$feature)
+
+        # retrieving the individual-level (random/varying) effects
+        participant_slopes <- coef(glmm)$participant
+        dimnames(participant_slopes)[[3]] <- sub("\\..*", "", dimnames(participant_slopes)[[3]])
+        participant_slopes <- data.frame(participant_slopes) |>
+            tibble::rownames_to_column(var = "participant")
+
+        # keeping only the columns that contains "value"
+        participant_slopes <- participant_slopes |> dplyr::select(.data$participant, tidyselect::contains("value") )
+
+        # reshaping into long format
+        kernels <- participant_slopes |>
+            tidyr::pivot_longer(
+                cols = -.data$participant,
+                names_to = c(".value", "feature"),
+                names_sep = "(?<=.)\\.(?=[^\\.]+$)" # regex to split on the last period
+                ) |>
+            dplyr::mutate(
+                kernel = dplyr::coalesce(.data$Estimate.value, .data$Estimate),
+                lower = dplyr::coalesce(.data$Q2.5.value, .data$Q2.5),
+                upper = dplyr::coalesce(.data$Q97.5.value, .data$Q97.5)
+                ) |>
+            dplyr::select(.data$participant, .data$feature, .data$kernel, .data$lower, .data$upper)
+
+        # appending fixed and random effects
+        kernels <- dplyr::bind_rows(kernels, group_kernel) |>
+            dplyr::mutate(feature = unique_feature_levels, .by = .data$participant)
+
+        # setting the class of the resulting object
+        class(kernels) <- c("glmm_kernel", "data.frame")
 
     }
 
-    # returning it
-    return (kernel)
+    # returning the kernels
+    return (kernels)
 
 }
 
@@ -312,6 +413,8 @@ plot.kernel <- function (
 
     # plotting the filters
     x |>
+        # avoiding log(0)
+        dplyr::filter(.data$feature > 0) |>
         ggplot2::ggplot(
             ggplot2::aes(
                 x = .data$feature,
@@ -332,7 +435,8 @@ plot.kernel <- function (
             inherit.aes = FALSE
             ) +
         # aesthetics
-        ggplot2::theme_bw(base_family = "Open Sans", base_size = 12) +
+        # ggplot2::theme_bw(base_family = "Open Sans", base_size = 12) +
+        ggplot2::theme_bw() +
         ggplot2::scale_colour_manual(
             values = MetBrewer::met.brewer(
                 name = "Johnson",
@@ -366,6 +470,7 @@ plot.kernel <- function (
 plot.glm_kernel <- function (x, log = TRUE, ...) {
 
     x |>
+        # avoiding log(0)
         dplyr::filter(.data$feature > 0) |>
         ggplot2::ggplot(
             ggplot2::aes(
@@ -385,7 +490,8 @@ plot.glm_kernel <- function (x, log = TRUE, ...) {
             y = "Kernel amplitude (a.u.)"
             ) +
         # aesthetics
-        ggplot2::theme_bw(base_family = "Open Sans", base_size = 12) +
+        # ggplot2::theme_bw(base_family = "Open Sans", base_size = 12) +
+        ggplot2::theme_bw() +
         ggplot2::scale_colour_manual(
             values = MetBrewer::met.brewer(
                 name = "Johnson",
@@ -406,11 +512,12 @@ plot.glm_kernel <- function (x, log = TRUE, ...) {
 plot.glmm_kernel <- function (x, log = TRUE, ...) {
 
     x |>
+        # avoiding log(0)
         dplyr::filter(.data$feature > 0) |>
         ggplot2::ggplot(
             ggplot2::aes(
-                x = as.numeric(.data$feature),
-                y = .data$kernel_gain,
+                x = .data$feature,
+                y = .data$kernel,
                 group = .data$participant,
                 color = .data$participant
                 )
@@ -422,9 +529,15 @@ plot.glmm_kernel <- function (x, log = TRUE, ...) {
             linewidth = 0.5, alpha = 0.5
             ) +
         # plotting the group-level average kernel
-        # plotting individual kernels
+        ggplot2::geom_ribbon(
+            data = x |> dplyr::filter(.data$participant == "group"),
+            ggplot2::aes(ymin = .data$lower, ymax = .data$upper, colour = NULL),
+            fill = "gray",
+            alpha = 0.5
+            ) +
         ggplot2::geom_line(
             data = x |> dplyr::filter(.data$participant == "group"),
+            colour = "black",
             linewidth = 1
             ) +
         ggplot2::labs(
@@ -432,7 +545,8 @@ plot.glmm_kernel <- function (x, log = TRUE, ...) {
             y = "Kernel amplitude (a.u.)"
             ) +
         # aesthetics
-        ggplot2::theme_bw(base_family = "Open Sans", base_size = 12) +
+        # ggplot2::theme_bw(base_family = "Open Sans", base_size = 12) +
+        ggplot2::theme_bw() +
         ggplot2::scale_colour_manual(
             values = MetBrewer::met.brewer(
                 name = "Johnson",